+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `DuplicatedImport` detector: PR [#381](https://github.com/nowarp/misti/pull/381)
 - `TransitiveImport` detector: Issue [#382](https://github.com/nowarp/misti/issues/382)
+- `SuboptimalSend`: Suggest replacing `self.forward` with `self.{notify,reply}`: Issue [#386](https://github.com/nowarp/misti/issues/386)
 
 ### Changed
 - `DuplicatedImport` and `TransitiveImport`: Don't check internal stdlib imports